requirements.txt
# Core
pytz==2023.3
Django==4.2.6
django-configurations==2.4.1
gunicorn==21.2.0
newrelic==9.1.1

# For the persistence stores
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
Markdown==3.5
django-filter==23.3

# Developer Tools
ipdb==0.13.13
ipython==8.16.1
mkdocs==1.5.3
flake8==6.1.0

# Testing
mock==5.1.0
factory-boy==3.3.0
coverage==7.3.2

# Static and Media Storage
django-storages==1.14.2
boto3==1.28.66

django-health-check==3.17.0

pytest==7.4.2
pytest-django==4.5.2